[scudo] Manage pages with MemMap in SizeClassAllocator64
Introduce a new data structure to manage the allocated pages from the
system. This is meant to deprecate certain memory system call wrappers
in Scudo, e.g., map()/unmap(). Besides, we would like to make
MapPlatformData to be appeared in platform specific data structure only.
Given that there are several allocators in Scudo and each of them has
different way of page management. The deprecation will be done in
several CLs. In this commit, we start from SizeClassAllocator64.

Reviewed By: cferris

Differential Revision: https://reviews.llvm.org/D146009

GitOrigin-RevId: 405ceaa09de04332384ab4e12a2ae5a79eb7d10d
Change-Id: Ie1619d1442980aae274478789df93f6990fbe74f