[scudo] Fix missing pushing 1 block to BatchClassId
This was happened rarely. The only case is when a thread is teared down
and it only has one block of BatchClass and the freelist of BatchClass
is empty. The impact is leaking 1 block of BatchClass and which is minor.

Differential Revision: https://reviews.llvm.org/D149141

Bug: b/279990828

GitOrigin-RevId: c266bfe278bb586ddb03111c6656d5fe3a2ad554
Change-Id: I62e425ba1fdc6a0575c301eae09ccc7289ce7ac6
Merged-In: I62e425ba1fdc6a0575c301eae09ccc7289ce7ac6
(cherry picked from commit 74fab6bece88074d9e4b007a8bbd5aed199e2037)