[scudo] Change secondary StatsAllocated update
In the secondary allocation routine, the StatsAllocated stat is
increased by BlockSize. However, in the deallocate routine, the
stat subtract uses CommitSize. CommitSize can be bigger than BlockSize
so this can lead to a negative calculated stat. Since the stats
are not guaranteed to be completely accurate, just add CommitSize
during allocation.

Reviewed By: Chia-hungDuan

Differential Revision: https://reviews.llvm.org/D150169

Bug: 281601348

GitOrigin-RevId: 526027df2d62432db3cfce3a61356369ed13cf5e
Change-Id: I257f0ccdbb8f53f6da9555f4c08a7329e066f5d7