[scudo] Make Options a reference for functions.
Modify all places that use the Options structure to be a const
reference. The underlying structure is a u32 so making it a
reference doesn't really do anything. However, if the structure
changes in the future it already works and avoids future coders
wondering why a structure is being passed by value. This also
makes it clear that the Options should not be modified in those functions.

Reviewed By: Chia-hungDuan

Differential Revision: https://reviews.llvm.org/D156372

GitOrigin-RevId: 867f2d9e5c9a736eb97fe5a51aa2850abfd02a96
Change-Id: Icc7ec50f09eec85e0102c701d76e1d5c0084d130