Add a clang format pre upload check.
Most of the time we are not uploading code directly into this project,
but we are going to start uploading some changes that are not going
to be upstreamed. Therefore, add the auto clang formatter and the
check for aosp first too.

The clang format file comes directly from the llvm project.

Test: Auto-formatted a CL.
Change-Id: I32e0be162f028ae65d4a30e6fad4e2a1a84581f9