Replace malloc_not_svelte with malloc_low_memory.
The malloc_not_svelte variable name is confusing and makes the
low memory config the default. Change this so that the default is
the regular allocator, and that malloc_low_memory is used to enable
the low memory allocator.

Update blueprint rules so that the default config is not the
low memory one.

Test: Verified scudo config is used by default.
Test: Verfified Android GO config uses the jemalloc low memory config.
Test: Build the scudo low memory config.
Change-Id: Ia3cddc97518cef0e5b86ff770d482d6e7a285712