[scudo] Apply the min release threshold to the group (#112014)
For the block smaller than a page size, one block is unlikely to
introduce more unused pages (at most 2 if it acrosses the page boundary
and both touched pages are unused). So it's better to apply the
threshold to reduce the time of scanning groups that can't release any
new pages.

GitOrigin-RevId: 53c9553562c778338a82574123fb7cfb943c4042
Change-Id: Ic4f601690e89d274ef33e75c836eb76da6d3a17a