[scudo] Skip special quarantine blocks in iterateOverChunks (#159892)
If a quarantine block is allocated, it will be passed through to a user
who calls iterateOverChunks. Make these special block allocations state
Quarantined so they are not passed to iterateOverChunks.

Remove the FIXME in the combined tests for quarantine since this fixes
those tests too.

Also add a specific new test that fails without this fix.

GitOrigin-RevId: d0eef22171096069fe4f6103b303bc5313c76fe7
Change-Id: I2114c3a216a255f10b8b22971839a6b3e938c227