[scudo] Only read urandom if getrandom syscall isn't available. (#161889)
If the getrandom system call is available, but the call returns an
error, it could mean that the system doesn't have enough randomness to
respond yet. Trying to read /dev/urandom will likely block and cause
initialization to be stalled. Therefore, return false in this case and
use the backup random data.

GitOrigin-RevId: f64223647e1969253ddc1bd8725bfacfc0456215
Change-Id: Icd5d4950269b8eb370e378f0034ad0ec492983c4