[scudo] Secondary release to OS uses LRU to scan. (#163691)
Before this change, the code would scan the entire set of cached entries
to find ones to be released. Now, it uses the LRUEntries list to iterate
over the live cached entries. In addition, remove the OldestTime
variable and replace it with OldestPresentEntry which will always be the
oldest entry in the LRU that has Time non-zero.

GitOrigin-RevId: dddcb84f152b99dfe7e117c02ab506c6c14b2f2b
Change-Id: I6f0c135589487289bfa36988b84b1f8807e23047