[scudo] Use setenv instead of putenv in ScudoCombinedTest.ZeroOnDeallocEnabledAndFlag (#173423)
This solves a stack-use-after-scope reported by AddressSanitizer within
the unsetenv call at end of the test, due to the "Options" buffer, that
we allocate on the stack, having already gone out of scope.

Unlike putenv, which stores the pointer to the passed string directly in
the environment, setenv creates an internal copy.

GitOrigin-RevId: 5c19f77a7e0c4b35c0efb511a7d9e2e436335e61
Change-Id: I2321e62bd0c803d330e12abab58d48d6b3439311